Changed parse_uint8 to use int pointers.
Rev: src/debug.c:1.13
Rev: src/disconnect.c:1.5
Rev: src/keyexchange.c:1.17
Rev: src/parse.c:1.13
Rev: src/parse.h:1.12
Rev: src/publickey_crypto.c:1.16

+++ src/debug.c
@@ -96,8 +96,8 @@ static int do_rec_debug(struct packet_handler *self,
 			struct lsh_string *packet)
 {
   struct simple_buffer buffer;
-  UINT8 msg_number;
-  UINT8 always_display;
+  int msg_number;
+  int always_display;
   UINT32 length;
   UINT8 *msg;
   int language;
