+++ src/publickey_crypto.c
@@ -540,7 +540,7 @@ int dh_process_client_msg(struct diffie_hellman_instance *self,
 			  struct lsh_string *packet)
 {
   struct simple_buffer buffer;
-  UINT8 msg_number;
+  int msg_number;
   
   simple_buffer_init(&buffer, packet->length, packet->data);
 
@@ -610,7 +610,7 @@ int dh_process_server_msg(struct diffie_hellman_instance *self,
 			  struct lsh_string *packet)
 {
   struct simple_buffer buffer;
-  UINT8 msg_number;
+  int msg_number;
   
   simple_buffer_init(&buffer, packet->length, packet->data);
